x_2/requirements.txt
Package                  Version   
------------------------ ----------
alabaster                0.7.12    
appdirs                  1.4.3     
appnope                  0.1.0     
astroid                  2.0.4     
Babel                    2.6.0     
backcall                 0.1.0     
beautifulsoup4           4.7.1     
biopython                1.73      
bioservices              1.6.0     
bleach                   3.0.2     
certifi                  2018.10.15
chardet                  3.0.4     
click                    6.7       
cloudpickle              0.8.0     
colorama                 0.4.1     
colorlog                 4.0.2     
colormap                 1.0.2     
cycler                   0.10.0    
decorator                4.3.0     
defusedxml               0.5.0     
docutils                 0.14      
easydev                  0.9.37    
entrypoints              0.2.3     
Flask                    1.0.2     
gevent                   1.4.0     
graphviz                 0.10.1    
greenlet                 0.4.15    
grequests                0.3.0     
idna                     2.8       
imagesize                1.1.0     
ipykernel                5.1.0     
ipython                  7.2.0     
ipython-genutils         0.2.0     
ipywidgets               7.4.2     
isort                    4.3.4     
itsdangerous             0.24      
jedi                     0.13.2    
Jinja2                   2.10      
jsonschema               2.6.0     
jupyter                  1.0.0     
jupyter-client           5.2.4     
jupyter-console          6.0.0     
jupyter-core             4.4.0     
keyring                  18.0.0    
kiwisolver               1.0.1     
lazy-object-proxy        1.3.1     
lxml                     4.3.0     
MarkupSafe               1.0       
matplotlib               3.0.2     
mccabe                   0.6.1     
meson                    0.51.0    
mistune                  0.8.4     
nbconvert                5.4.0     
nbformat                 4.4.0     
notebook                 5.7.4     
numpy                    1.15.4    
numpydoc                 0.8.0     
packaging                18.0      
pandas                   0.23.4    
pandocfilters            1.4.2     
parso                    0.3.1     
pexpect                  4.6.0     
pickleshare              0.7.5     
pip                      19.0.3    
pipenv                   2018.11.14
plotly                   3.6.1     
prometheus-client        0.5.0     
prompt-toolkit           2.0.7     
psutil                   5.6.0     
ptyprocess               0.6.0     
pycodestyle              2.5.0     
pyflakes                 2.1.1     
Pygments                 2.3.1     
pylint                   2.1.1     
pylint-django            2.0.9     
pylint-plugin-utils      0.5       
pyparsing                2.3.0     
PyQt5                    5.8       
PyQt5-sip                4.19.14   
pysam                    0.15.2    
python-dateutil          2.7.5     
pytz                     2018.9    
PyYAML                   3.13      
pyzmq                    17.1.2    
QtAwesome                0.5.7     
qtconsole                4.4.3     
QtPy                     1.6.0     
requests                 2.21.0    
requests-cache           0.4.13    
retrying                 1.3.3     
rope                     0.12.0    
scipy                    1.2.0     
Send2Trash               1.5.0     
setuptools               40.8.0    
sip                      4.19.8    
six                      1.11.0    
slackclient              1.3.0     
snowballstemmer          1.2.1     
soupsieve                1.7       
Sphinx                   1.8.3     
sphinxcontrib-websupport 1.1.0     
spyder-kernels           0.4.2     
suds-jurko               0.6       
terminado                0.8.1     
testpath                 0.4.2     
tornado                  5.1.1     
traitlets                4.3.2     
urllib3                  1.24.1    
virtualenv               16.1.0    
virtualenv-clone         0.4.0     
wcwidth                  0.1.7     
webencodings             0.5.1     
websocket-client         0.54.0    
Werkzeug                 0.14.1    
wheel                    0.33.1    
widgetsnbextension       3.4.2     
wrapt                    1.10.11   
wurlitzer                1.0.2     
xmltodict                0.11.0    
